services/instagram/instagrapi/bots/afadesign.official/scraper/test_instagram_scraper.py
# test_instagram_scraper.py
import asyncio
import logging

import aiohttp
import aioredis
import pytest

from instagrapi import Client


@pytest.fixture
async def redis():
    redis_url = "redis://localhost:6379"
    redis = await aioredis.create_redis_pool(redis_url)
    yield redis
    redis.close()
    await redis.wait_closed()


class UserCacheManager:
    pass


@pytest.fixture
async def user_cache_manager(redis):
    logging.getLogger("UserCacheManager")
    yield UserCacheManager()


class FollowerCacheManager:
    pass


@pytest.fixture
async def follower_cache_manager(redis):
    logging.getLogger("FollowerCacheManager")
    yield FollowerCacheManager()


class ConfigManager:
    pass


class InstagramAPI:
    pass


@pytest.fixture
async def instagram_api(user_cache_manager, follower_cache_manager):
    logging.getLogger("InstagramAPI")
    cl = Client()
    cl.session = aiohttp.ClientSession()
    ConfigManager()
    yield InstagramAPI()


class CacheInvalidationManager:
    pass


@pytest.fixture
async def cache_invalidation_manager(redis):
    logging.getLogger("CacheInvalidationManager")
    yield CacheInvalidationManager()


class HTTPClient:
    pass


@pytest.fixture
async def http_client():
    logging.getLogger("HTTPClient")
    yield HTTPClient()


# Add more fixtures if needed


@pytest.mark.asyncio
async def test_login_user(instagram_api):
    cl = Client()
    await instagram_api.login_user(cl)
    # Add assertions to check if login was successful


@pytest.mark.asyncio
async def test_refresh_session_if_needed(instagram_api):
    cl = Client()
    await instagram_api.login_user(cl)
    await instagram_api.refresh_session_if_needed(cl)
    # Add assertions to check if session was refreshed if needed


# Add more tests for other methods of InstagramAPI


@pytest.mark.asyncio
async def test_cache_user_data(user_cache_manager):
    user_id = "test_user_id"
    user_data = {"name": "Test User", "scraper": 1000}
    await user_cache_manager.cache_user_data(user_id, user_data)
    cached_data = await user_cache_manager.get_cached_user_data(user_id)
    assert cached_data == user_data


# Add more tests for other methods of UserCacheManager


@pytest.mark.asyncio
async def test_cache_followers(follower_cache_manager):
    user_id = "test_user_id"
    followers = [{"username": "follower1"}, {"username": "follower2"}]
    await follower_cache_manager.cache_followers(user_id, followers)
    cached_followers = await follower_cache_manager.get_cached_followers(user_id)
    assert cached_followers == followers


# Add more tests for other methods of FollowerCacheManager


@pytest.mark.asyncio
async def test_start_cache_invalidation_listener(
    cache_invalidation_manager, redis, caplog
):
    caplog.set_level(logging.INFO)
    await asyncio.create_task(
        cache_invalidation_manager.start_cache_invalidation_listener()
    )
    await asyncio.sleep(0.1)  # Wait for listener to start
    await redis.publish("cache_invalidation_channel", "test_cache:invalidation_key")
    await asyncio.sleep(0.1)  # Wait for message to be processed
    assert "Received cache invalidation message" in caplog.text
    assert "Invalidating cache entry" in caplog.text


# Add more tests for other methods of CacheInvalidationManager


@pytest.mark.asyncio
async def test_fetch_data(http_client, mocker):
    url = "http://test.com/api/data"
    mocker.patch.object(http_client, "handle_response", return_value={"key": "value"})
    response = await http_client.fetch_data(url)
    assert response == {"key": "value"}


# Add more tests for other methods of HTTPClient
